requirements.txt
python-dotenv
py-cord
GitPython
python-Levenshtein

dnspython==1.16.0
PyNaCl==1.3.0